+++ plone/outputfilters/filters/resolveuid_and_caption.py
@@ -7,6 +7,7 @@ from DocumentTemplate.DT_Util import html_quote
 from DocumentTemplate.DT_Var import newline_to_br
 from plone.outputfilters.browser.resolveuid import uuidToObject
 from plone.outputfilters.interfaces import IFilter
+from plone.registry.interfaces import IRegistry
 from Products.CMFCore.interfaces import IContentish
 from sgmllib import SGMLParseError
 from sgmllib import SGMLParser
@@ -19,6 +20,7 @@ from zExceptions import NotFound
 from ZODB.POSException import ConflictError
 from zope.cachedescriptors.property import Lazy as lazy_property
 from zope.component import getAllUtilitiesRegisteredFor
+from zope.component import getUtility
 from zope.component.hooks import getSite
 from zope.interface import Attribute
 from zope.interface import implementer
@@ -48,6 +50,18 @@ class IImageCaptioningEnabler(Interface):
 class IResolveUidsEnabler(Interface):
     available = Attribute(
         "Boolean indicating whether UID links should be resolved.")
+
+
+@implementer(IImageCaptioningEnabler)
+class ImageCaptioningEnabler(object):
+
+    @property
+    def available(self):
+        name = 'plone.image_captioning'
+        registry = getUtility(IRegistry)
+        if name in registry:
+            return registry[name]
+        return False
 
 
 @implementer(IResolveUidsEnabler)
@@ -275,9 +289,9 @@ class ResolveUIDAndCaptionFilter(SGMLParser):
         klass = attributes['class']
         del attributes['class']
         del attributes['src']
-        if 'width' in attributes:
+        if 'width' in attributes and attributes['width']:
             attributes['width'] = int(attributes['width'])
-        if 'height' in attributes:
+        if 'height' in attributes and attributes['height']:
             attributes['height'] = int(attributes['height'])
         view = fullimage.unrestrictedTraverse('@@images', None)
         if view is not None: